==> code/04_Generate_3DPlots.ipynb <==
import pandas as pd
import numpy as np
import pickle
import time
import matplotlib.colors as clr
import seaborn as sns
import plotly.graph_objs as go
from plotly.offline import download_plotlyjs, init_notebook_mode, plot, iplot, plot_mpl
init_notebook_mode(connected=True)
import matplotlib.pyplot as plt
from rdkit.Chem.Draw import IPythonConsole
from rdkit.Chem import PandasTools
from rdkit.Chem import Draw

Giulia Menichetti menicgiulia@gmail.com

# Import Data for Experimental Outcomes

specifics='Jaccard_Binary_8192b_R3'
#viz_data = pd.read_pickle('../data/Broad_Outcomes_Keys_with_fingerprint_embedding'+specifics+'.pkl')
viz_data = pd.read_csv('../data/data3Dspace.csv')

viz_data

# Colormap for Experimental Outcomes

# colormap
cmapdict={'No-Effect': (179/255,226/255,205/255), 
'Low (25%)': (102/255,194/255,165/255),
'Very Weak (35%)': (27/255,158/255,119/255),          
'Weak (50%)': (231/255,41/255,138/255),          
'Strong (80%)': (117/255,112/255,179/255),          
'Cyto':(217/255,95/255,2/255)
}
ordered_classes = list(cmapdict.keys())


# colormap for Plotly
cmapdictplotly={k:'rgb'+str(tuple([v*255 for v in cmapdict[k]])) for k in ordered_classes}
alphadict={'No-Effect':0.1, 
'Cyto':0.3,
'Low (25%)': 0.3,
'Very Weak (35%)': 0.3,          
'Weak (50%)': 0.8,   
'Strong (80%)':0.8}

# HTML file Drug 3D Space with Experimental Classes

def make_label(row, name='pert_iname'):
    return '<br>Name: ' + row[name] \
        + '<br>Outcome: ' + row['Outcome'] \
        + '<br>BroadId: ' + row['truncated_broadid'] \
        + '<br>Smile: ' + row['rdkit_smiles'] \
        + '<br>InChiKey: ' + row['rdkit_InChIKey']

traces = []


for c in ordered_classes:
    traces.append( go.Scatter3d(
        x=viz_data[viz_data['Outcome'] == c].coord1.tolist(),
        y=viz_data[viz_data['Outcome'] == c].coord2.tolist(),
        z=viz_data[viz_data['Outcome'] == c].coord3.tolist(),
        mode='markers',
        marker=dict(
            opacity=alphadict[c],
            size=3,
            color=cmapdictplotly[c]
        ),
        name=c,
        text=viz_data[viz_data['Outcome'] == c].apply(make_label, axis=1),
        hoverinfo='text'
    ))
    
layout = go.Layout(
    title=specifics
)

fig = go.Figure(data=traces, layout=layout)
plot(fig, filename = '../figs/'+specifics+'.html', auto_open=True)

# Import Data for ATC Codes

drugclassification=pd.read_csv('../data/drug-enrichment-Strong (80%) Weak (50%)-for3Dspace.tsv', sep='\t')

# Colormap for ATC codes

# colormap drug classes
listcolors=['rgb'+str(tuple([k*255 for k in n])) for n in (plt.get_cmap('Accent').colors[:-1]+plt.get_cmap('Dark2').colors[:-1]+plt.get_cmap('Set1').colors[:-1]+plt.get_cmap('Set2').colors[:-1])]
cmapdictplotlyDC={drugclassification['Chemical Group'][k]: listcolors[k] for k in range(len(drugclassification['Chemical Group']))}

# HTML file Drug 3D Space with enriched ATC codes

specifics='Enriched_Drug_Classes'
traces = []

# Strong and Weak Drugs
traces.append( go.Scatter3d(
    x=viz_data[viz_data['Outcome'].isin(['Strong (80%)', 'Weak (50%)'])].coord1.tolist(),
    y=viz_data[viz_data['Outcome'].isin(['Strong (80%)', 'Weak (50%)'])].coord2.tolist(),
    z=viz_data[viz_data['Outcome'].isin(['Strong (80%)', 'Weak (50%)'])].coord3.tolist(),
    mode='markers',
    marker=dict(
    opacity=0.3,
    size=3,
    color='rgb(72.0, 72.0, 72.0)'
        ),
    name="Strong (80%) & Weak (50%)",
    text=viz_data[viz_data['Outcome'].isin(['Strong (80%)', 'Weak (50%)'])].apply(make_label, axis=1),
    hoverinfo='text'
    ))


count=0
for c in drugclassification['Chemical Group']:
    dfsel=viz_data[viz_data['truncated_broadid'].isin(drugclassification['Broad IDs'].iloc[count].split(','))]
    
    print('%%%%%%%%%%%%%%%%%%')
    print(c)
    print('Number of Drugs')
    print(dfsel.shape[0])
    print(drugclassification['Broad IDs'].iloc[count].split(','))
    print('%%%%%%%%%%%%%%%%%%')
    
    count+=1
    
    traces.append( go.Scatter3d(
        x=dfsel.coord1.tolist(),
        y=dfsel.coord2.tolist(),
        z=dfsel.coord3.tolist(),
        mode='markers',
        marker=dict(
            opacity=0.3,
            size=10,
            color=cmapdictplotlyDC[c]
        ),
        name=c,
        text=dfsel.apply(make_label, axis=1),
        hoverinfo='text'
    ))
    
layout = go.Layout(
    title=specifics
)

fig = go.Figure(data=traces, layout=layout)
plot(fig, filename = '../figs/'+specifics+'.html', auto_open=True)
